+++ common/src/main/java/dev/terminalmc/clientsort/network/handler/PayloadHandler.java
@@ -20,6 +20,7 @@ import dev.terminalmc.clientsort.ClientSort;
 import dev.terminalmc.clientsort.exception.PayloadHandlerException;
 import dev.terminalmc.clientsort.exception.PayloadHandlerException.InconsistentStateException;
 import dev.terminalmc.clientsort.exception.PayloadHandlerException.InvalidDataException;
+import dev.terminalmc.clientsort.mixin.accessor.ContainerAccessor;
 import dev.terminalmc.clientsort.network.handler.validate.PayloadResult;
 import dev.terminalmc.clientsort.platform.services.PlatformServices;
 import net.minecraft.network.protocol.common.custom.CustomPacketPayload;
@@ -98,6 +99,8 @@ public abstract class PayloadHandler {
             if (menu != null) {
                 menu.resumeRemoteUpdates();
                 menu.broadcastChanges();
+                ((ContainerAccessor) menu.slots.getFirst().container).clientsort$setChanged();
+                ((ContainerAccessor) menu.slots.getLast().container).clientsort$setChanged();
             }
             if (PlatformServices.getInstance().canSendToPlayer(player, responseType)) {
                 PlatformServices.getInstance()